fix infinite recursion
a method which iterates through the child elements of a given
subtree root (which may be named different things depending on the
element) is called recursively.

the postwalk merging process is very slow and needs to be optimized

also, methods which return integers now return nan when there is a zero
division error. more correctly, it would return +/- infinity if the
numerator is nonzero and nan if the numerator were zero